+++ functions/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "firebase-admin": "~6.0.0",
     "firebase-functions": "^2.0.3",
+    "glob": "^7.1.3",
     "sib-api-v3-sdk": "^7.0.3"
   },
   "devDependencies": {
