--- functions/src/importMp3IntoSQR.js
@@ -1,8 +1,0 @@
-
-exports.handleNewUploads = (object, db, storeFileNameToDB) => {
-    const filePath = object.name;
-    
-    storeFileNameToDB(filePath, db);
-
-    return Promise.resolve(1);
-}
